makes relationship between teams and events (Ilias)
