+++ models/event.js
@@ -1,19 +1,19 @@
 var mongoose = require('mongoose');
 var Schema = mongoose.Schema;
-// var Team = require('../models/team');
+var Team = require('../models/team');
 
 var EventSchema = new Schema ({
-  homeTeam: String,
-  awayTeam: String,
+  // homeTeam: String,
+  // awayTeam: String,
   // trying to ref teams in events
-  // homeTeam: [{
-  //   type: Schema.Types.ObjectId,
-  //   ref: 'Team'
-  // }],
-  // awayTeam: [{
-  //   type: Schema.Types.ObjectId,
-  //   ref: 'Team'
-  // }]
+  homeTeam: [{
+    type: Schema.Types.ObjectId,
+    ref: 'Team'
+  }],
+  awayTeam: [{
+    type: Schema.Types.ObjectId,
+    ref: 'Team'
+  }],
   date: Date,
   time: String,
   createdAt: { type : Date, default: Date.now() }
